html/browsers/browsing-the-web/navigating-across-documents/014.html

<!doctype html>
<title> Link with javascript onclick form submission script order </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:parent.events.push('submit');parent.resolve();"></form>
<a target="test" onclick="document.forms[0].submit()">Test</a>
<script>

promise_test(async function(t) {
  window.promise = new Promise(resolve => window.resolve = resolve);

  document.getElementsByTagName("a")[0].click();
  window.events = [];
  window.events.push('after script');

  await promise;
  assert_array_equals(events, ['after script', 'submit']);
  t.done();
});
</script>
