Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/014-expected.txt


FAIL  Link with javascript onclick form submission script order  assert_array_equals: property 0, expected "after script" but got "submit"

Test
